+++ modules/public_pages/erp/ledger/general_ledger/models/GLTransaction.php
@@ -1012,6 +1012,55 @@ class GLTransaction extends DataObject
         return $eu_tax_elements;
     }
 
+    /**
+     * Handle tax for PVA and Reverse Charge
+     *
+     * @param $transaction
+     * @param $type  'RC' = Reverse Charge, 'PVA' = Postponed VAT Accounting
+     * @param array $errors
+     * @return void
+     */
+    public static function makeTax($transaction, $tax_type='none',  &$errors = array())
+    {
+        $mult = self::$multipliers[$transaction['source']][$transaction['type']];
+        $eu_tax_element = $transaction;
+
+        $glparams = DataObjectFactory::Factory('GLParams');
+
+        switch ($tax_type) {
+            case 'PVA':
+                $eu_tax_element['glaccount_id'] = $glparams->vat_postponed_account();
+                break;
+            case 'RC':
+                $eu_tax_element['glaccount_id'] = $glparams->vat_reverse_charge_account();
+                break;
+            default:
+                $errors[] = "Tax type not specified";
+        }
+
+        if ($eu_tax_element['glaccount_id'] === false) {
+            $errors[] = "Account Code for tax '{$tax_type}' not found";
+        }
+
+        $eu_tax_element['glcentre_id'] = $glparams->balance_sheet_cost_centre();
+
+        if ($eu_tax_element['glcentre_id'] === false) {
+            $errors[] = 'Balance Sheet Cost Centre Code not found';
+        }
+
+        $eu_tax_element['value'] = bcmul($mult, $eu_tax_element['value']);
+        $eu_tax_element['twinvalue'] = bcmul($mult, $eu_tax_element['twinvalue']);
+
+        $eu_tax_elements = array();
+        $eu_tax_elements[] = GLTransaction::Factory($eu_tax_element, $errors);
+
+        $eu_tax_element['value'] *= - 1;
+        $eu_tax_element['twinvalue'] *= - 1;
+        $eu_tax_elements[] = GLTransaction::Factory($eu_tax_element, $errors);
+
+        return $eu_tax_elements;
+    }
+
     public static function makeCBLine($transaction, &$errors = array())
     {
         $mult = self::$multipliers[$transaction['source']][$transaction['type']];